refactorización de lógica de negocio para el módulo de usuarios

+++ src/api/users/users.service.ts
@@ -1,6 +1,13 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
+import * as bcrypt from 'bcrypt';
 
 import { User } from './entities/user.entity';
+import { Status } from '../statuses/entities/status.entity';
+import { Role } from '../roles/entities/role.entity';
 import { UsersRepository } from './users.repository';
 import { StatusesRepository } from '../statuses/statuses.repository';
 import { RolesRepository } from '../roles/roles.repository';
@@ -22,13 +29,34 @@ export class UsersService extends FilterService<User, FilterUsersDto> {
   }
 
   async create(createUserDto: CreateUserDto): Promise<User> {
-    const newUser = this.usersRepository.create(createUserDto);
+    const salt = await bcrypt.genSalt(10);
+    createUserDto.password = await bcrypt.hash(createUserDto.password, salt);
 
-    newUser.status = await this.statusesRepository.findOne({
-      where: { abbreviation: StatusAbbreviations.GEN_ACTIVE_STATUS },
-    });
+    const userByEmail = await this.usersRepository.findOneByEmail(
+      createUserDto.email,
+    );
 
-    newUser.roles = await this.rolesRepository.findByIds(createUserDto.roles);
+    if (userByEmail) {
+      throw new BadRequestException(
+        `The email given (${createUserDto.email}) is alredy exist`,
+      );
+    }
+
+    const roles = await this.rolesRepository.findByIds(createUserDto.roles);
+
+    if (createUserDto.roles.length !== roles.length) {
+      throw new BadRequestException(
+        `One of given roles (${createUserDto.roles.toString()}) does not exist in the application`,
+      );
+    }
+
+    const newUser = this.usersRepository.create({ ...createUserDto, roles });
+
+    newUser.status = await this.statusesRepository.findOneByAbbreviation(
+      StatusAbbreviations.GEN_ACTIVE_STATUS,
+    );
+
+    newUser.roles = roles;
 
     return await this.usersRepository.save(newUser);
   }
@@ -37,9 +65,9 @@ export class UsersService extends FilterService<User, FilterUsersDto> {
     filterUsersDto: FilterUsersDto,
   ): Promise<PaginatedResourceResultDto<User>> {
     const filterOptions = this.filterByParams(filterUsersDto);
-    const usersResult = await this.usersRepository.findAndCount(filterOptions);
-    const users = usersResult[0];
-    const count = usersResult[1];
+    const [users, count] = await this.usersRepository.findAndCount(
+      filterOptions,
+    );
     const page = filterUsersDto.page ?? 1;
     const limit = Math.ceil(count / filterOptions.take);
 
@@ -65,26 +93,59 @@ export class UsersService extends FilterService<User, FilterUsersDto> {
 
   async update(id: number, updateUserDto: UpdateUserDto): Promise<User> {
     const user = await this.usersRepository.findOne(id);
+    const { roles, status, email } = updateUserDto;
+    let findStatus: Status | undefined;
+    let findRoles: Role[] | undefined;
 
-    if (updateUserDto.status) {
-      const status = await this.statusesRepository.findOne(
-        updateUserDto.status,
-      );
-      user.status = status;
+    if (email) {
+      const user = await this.usersRepository.findOneByEmail(email);
+
+      if (user) {
+        throw new BadRequestException(
+          `The email given (${email}) is alredy exist`,
+        );
+      }
     }
 
-    if (updateUserDto.roles) {
-      const roles = await this.rolesRepository.findByIds(updateUserDto.roles);
-      user.roles = roles;
+    if (status) {
+      findStatus = await this.statusesRepository.findOne(status);
+
+      if (!findStatus) {
+        throw new BadRequestException(
+          `Status (${updateUserDto.status}) does not exist in the application`,
+        );
+      }
+
+      user.status = findStatus;
     }
 
-    this.usersRepository.merge(user, updateUserDto);
+    if (roles) {
+      findRoles = await this.rolesRepository.findByIds(roles);
 
-    return this.usersRepository.save(user);
+      if (updateUserDto.roles.length !== roles.length) {
+        throw new BadRequestException(
+          `One of given roles (${updateUserDto.roles.toString()}) does not exist in the application`,
+        );
+      }
+
+      user.roles = findRoles;
+    }
+
+    const userUpdated = this.usersRepository.merge(user, {
+      ...updateUserDto,
+      roles: findRoles,
+      status: findStatus,
+    });
+
+    return this.usersRepository.save(userUpdated);
   }
 
   async remove(id: number): Promise<User> {
     const user = await this.usersRepository.findOne(id);
+
+    if (!user) {
+      throw new NotFoundException(`User #${id} not found`);
+    }
 
     if (await this.usersRepository.delete(id)) {
       return user;
